fix dead code and handle stream seek underflow in CliprdrStream_Seek
##Summary
Fixes a dead code issue and boundary check in CliprdrStream_Seek where negative stream offset calculations were not correctly evaluated due to newoffset being an unsigned integer(ULONGLONG)

##Problem Analysis
In CliprdrStream_Seek, newoffset is declared as ULONGLONG(unsigned 64-bit integer), while the input displacement dlibMove.QuadPart is a signed 64-bit integer(LONGLONG).

1. Dead Code: The validation check if (newoffset < 0) checks if an unsigned variable is less than zero. Since ULONGLONG can never be negative, this condition is optimized away by the compiler or constantly evaluates to FALSE.

2. Underflow Logic Bypass: When a negative offset is passed in STREAM_SEEK_CUR or STREAM_SEEK_END that results in an invalid negative stream position, newoffset underflows into a large positive integer. Because newoffset < 0 is bypassed, the code relies purely on the underflow wrapping around to trigger the newoffset >= size condition.

##Proposed Fix
Cast newoffset to (LONGLONG) in the negative boundary check: if ((LONGLONG)newoffset < 0 || newoffset >= instance->m_lSize.QuadPart).

This ensures that any underflown bit-pattern resulting from a negative position calculation is correctly re-interpreted as a signed negative value and rejected with E_FAIL.

Keeps the original variable declarations intact to prevent unnecessary code churn.


##Question for Maintainers:

I noticed that newoffset is declared as ULONGLONG, which causes the newoffset < 0 condition to be bypassed. I added an explicit (LONGLONG) cast to restore the negative bound check while keeping the code change minimal.

Please let me know if there was another intended approach for handling negative offsets here, or if this change aligns with your expectations!

+++ client/Windows/wf_cliprdr.c
@@ -300,7 +300,7 @@ static HRESULT STDMETHODCALLTYPE CliprdrStream_Seek(IStream* This, LARGE_INTEGER
 			return E_INVALIDARG;
 	}
 
-	if (newoffset < 0 || newoffset >= instance->m_lSize.QuadPart)
+	if (LONGLONG)newoffset < 0 || newoffset >= instance->m_lSize.QuadPart)
 		return E_FAIL;
 
 	instance->m_lOffset.QuadPart = newoffset;
